requirements.txt
anthropic==0.8.1
python-dotenv==1.0.0
pandas==2.1.1
tqdm==4.66.1 